split into 2 goroutines to handle SDL window and CHIP8 emulation separately

+++ chip8/config.go
@@ -2,14 +2,14 @@ package chip8
 
 // Config represents the configuration for the CHIP8 machine
 type Config struct {
-	ResolutionX, ResolutionY int // num pixels
-	SizeProgram              int // bytes
-	SizeStack                int // bytes
-	SizeDisplay              int // bytes
-	NumRegisters             int // num 16-bit registers
-	ClockFreq                int // Hz
-	ScreenRefreshFreq        int // Hz
-	TimerDecrementFreq       int // Hz
+	ResolutionX, ResolutionY int     // num pixels
+	SizeProgram              int     // bytes
+	SizeStack                int     // bytes
+	SizeDisplay              int     // bytes
+	NumRegisters             int     // num 16-bit registers
+	ClockFreq                float32 // Hz
+	ScreenRefreshFreq        int     // Hz
+	TimerDecrementFreq       int     // Hz
 }
 
 // GetDefaultConfig returns the default CHIP8 configuration
